devices/video/decoder: move set_output_buffer_count logic to vda
The VDA requires the maximum number of output buffers to be specified
after a resolution change event, and before the first buffer is
imported. This constraint is specific to VDA and does not necessarily
apply to other backends, so move it into the VDA device module.

This allows us to simplify the DecoderSession interface a bit.

BUG=b:169295147
BUG=b:161774071
TEST=Youtube video playback works on kukui-arc-r.
TEST=Youtube video playback works on hatch-arc-r.

Change-Id: I4fb3d60421a2cd625b5b173000f65d0b8b8bf609